refactor: move ANSIEscapeCode and tests to own module
